prescription download, field dropdown component, ui fixes
