+++ src/components/FileUpload/index.jsx
@@ -1,5 +1,5 @@
 /* eslint-disable no-alert */
-import React, { useState, useEffect, useMemo } from 'react';
+import React, { useState, useEffect } from 'react';
 import { useAuth0 } from '@auth0/auth0-react';
 import { BlobServiceClient } from '@azure/storage-blob';
 import {
@@ -17,31 +17,26 @@ import {
 } from '@mui/material';
 import CheckCircleIcon from '@mui/icons-material/CheckCircle';
 import { PSAButton, PSATextField } from 'shared-react-components/src';
-import axios from 'axios';
 import shpjs from 'shpjs';
+import { useDispatch, useSelector } from 'react-redux';
 import { isValidGeoJSON } from '../../utils/geojsonUtils';
-import { azureSASToken, containerName, ncalcApiUrl, storageAccountName } from '../../utils/keys';
+import { azureSASToken, containerName, storageAccountName } from '../../utils/keys';
+import FieldDropdown from '../../shared/FieldDropdown/FieldDropdown';
+import { get, set } from '../../store/redux-autosetters';
 
-const API_BASE_URL = ncalcApiUrl;
 const MAP_TYPES = ['Spray Map', 'Yield Map'];
 const ROLES = ['NIFA-Soy', 'Willard', 'Growmark'];
 
 const UploadMap = () => {
-  const {
-    user, isAuthenticated, isLoading, getAccessTokenSilently,
-  } = useAuth0();
+  const { user, isAuthenticated, isLoading } = useAuth0();
 
+  const dispatch = useDispatch();
   const matchesMd = useMediaQuery((theme) => theme.breakpoints.down('md'));
-
-  const [fieldOptions, setFieldOptions] = useState([]);
-  const [isFetching, setIsFetching] = useState(null);
 
   const [existingMapTypes, setExistingMapTypes] = useState(new Set());
   const [checkingMapTypes, setCheckingMapTypes] = useState(false);
 
-  const [filterProgram, setFilterProgram] = useState(null);
-  const [filterGrower, setFilterGrower] = useState(null);
-  const [selectedField, setSelectedField] = useState(null);
+  const selectedField = useSelector(get.selectedField);
   const [selectedSeason, setSelectedSeason] = useState(null);
   const [mapType, setMapType] = useState(MAP_TYPES[0]);
   const [selectedFile, setSelectedFile] = useState(null);
@@ -50,6 +45,10 @@ const UploadMap = () => {
 
   const roles = user?.['https://dst-ncalc.org/claims'] || [];
   const isAllowed = isAuthenticated && (roles.includes('admin') || roles.some((r) => ROLES.includes(r)));
+
+  useEffect(() => {
+    setSelectedSeason(null);
+  }, [selectedField]);
 
   useEffect(() => {
     const checkExistingMaps = async () => {
@@ -97,75 +96,6 @@ const UploadMap = () => {
 
     checkExistingMaps();
   }, [selectedField, selectedSeason, isAuthenticated]);
-
-  // Fetch All Fields
-  useEffect(() => {
-    const fetchFields = async () => {
-      try {
-        setIsFetching(true);
-        const token = await getAccessTokenSilently();
-        const response = await axios.get(`${API_BASE_URL}/fields-identifiers`, {
-          headers: { Authorization: `Bearer ${token}` },
-        });
-        setFieldOptions(response.data);
-      } catch (error) {
-        console.error('Failed to load options', error);
-      } finally {
-        setIsFetching(false);
-      }
-    };
-
-    if (isAuthenticated) {
-      fetchFields();
-    }
-  }, [isAuthenticated, getAccessTokenSilently]);
-
-  // Returns list of case insensitive unique strings
-  const getUniqueCaseInsensitive = (list) => {
-    const seen = new Set();
-    return list.filter((item) => {
-      if (!item) return false;
-      const lower = item.toLowerCase();
-      if (seen.has(lower)) return false;
-      seen.add(lower);
-      return true;
-    }).sort();
-  };
-
-  const uniquePrograms = useMemo(() => {
-    const programs = fieldOptions.map((f) => f.properties.programName);
-    return getUniqueCaseInsensitive(programs);
-  }, [fieldOptions]);
-
-  const uniqueGrowers = useMemo(() => {
-    let filtered = fieldOptions;
-    if (filterProgram) {
-      filtered = filtered.filter((f) => f.properties.programName?.toLowerCase() === filterProgram.toLowerCase());
-    }
-    const growers = filtered.map((f) => f.properties.growerName);
-    return getUniqueCaseInsensitive(growers);
-  }, [fieldOptions, filterProgram]);
-
-  const filteredFieldList = useMemo(() => fieldOptions.filter((f) => {
-    const pName = f.properties.programName;
-    const gName = f.properties.growerName;
-
-    const matchProgram = !filterProgram || (pName && pName.toLowerCase() === filterProgram.toLowerCase());
-    const matchGrower = !filterGrower || (gName && gName.toLowerCase() === filterGrower.toLowerCase());
-
-    return matchProgram && matchGrower;
-  }), [fieldOptions, filterProgram, filterGrower]);
-
-  const handleProgramFilterChange = (newVal) => {
-    setFilterProgram(newVal);
-    setFilterGrower(null);
-    setSelectedField(null);
-  };
-
-  const handleGrowerFilterChange = (newVal) => {
-    setFilterGrower(newVal);
-    setSelectedField(null);
-  };
 
   const handleUpload = async () => {
     if (!selectedFile || !selectedField || !selectedSeason) {
@@ -261,7 +191,7 @@ const UploadMap = () => {
       // Reset file input
       setSelectedFile(null);
       setUploadProgress(0);
-      setSelectedField(null);
+      dispatch(set.selectedField(null));
       setMapType(MAP_TYPES[0]);
     } catch (error) {
       alert(`Upload failed: ${error.message || 'Unknown error occurred'}`);
@@ -381,79 +311,7 @@ const UploadMap = () => {
             Use the filters below to find your field.
           </Typography>
 
-          <Grid container>
-            <Grid item xs={12} md={6} sx={{ pr: { md: 1 }, pb: { xs: 2 } }}>
-              <Autocomplete
-                options={uniquePrograms}
-                value={filterProgram}
-                onChange={(e, val) => handleProgramFilterChange(val)}
-                renderInput={(params) => <PSATextField {...params} label="Filter by Program" />}
-              />
-            </Grid>
-            <Grid item xs={12} md={6} sx={{ pl: { md: 1 } }}>
-              <Autocomplete
-                options={uniqueGrowers}
-                value={filterGrower}
-                onChange={(e, val) => handleGrowerFilterChange(val)}
-                disabled={!filterProgram && uniqueGrowers.length > 50}
-                renderInput={(params) => <PSATextField {...params} label="Filter by Grower" />}
-              />
-            </Grid>
-          </Grid>
-
-          <Typography variant="body1" color="textSecondary">
-            Select the field details below to ensure the file is saved in the correct folder.
-          </Typography>
-
-          <Autocomplete
-            loading={isFetching}
-            loadingText="Loading fields..."
-            options={filteredFieldList}
-            value={selectedField}
-            key={`${filterProgram}-${filterGrower}`}
-            onChange={(event, newValue) => {
-              setSelectedField(newValue);
-              setSelectedSeason(null);
-            }}
-            getOptionLabel={(option) => {
-              const p = option.properties;
-              return `${p.programName} / ${p.growerName} / ${p.farmName} / ${p.fieldName}`;
-            }}
-            renderOption={(props, option) => (
-              <Box component="li" {...props}>
-                <Stack>
-                  <Typography variant="subtitle1" sx={{ fontWeight: 'bold' }}>
-                    {option.properties.fieldName}
-                  </Typography>
-                  <Typography variant="caption" color="textSecondary">
-                    {option.properties.programName}
-                    {' - '}
-                    {option.properties.growerName}
-                    {' - '}
-                    {option.properties.farmName}
-                    {/* {' - '}
-                    {option.properties.season} */}
-                  </Typography>
-                </Stack>
-              </Box>
-            )}
-            renderInput={(params) => (
-              <PSATextField
-                {...params}
-                label="Select Field (Program / Grower / Farm / Field)"
-                placeholder="Type to search..."
-                InputProps={{
-                  ...params.InputProps,
-                  endAdornment: (
-                    <>
-                      {isFetching ? <CircularProgress color="inherit" size={20} /> : null}
-                      {params.InputProps.endAdornment}
-                    </>
-                  ),
-                }}
-              />
-            )}
-          />
+          <FieldDropdown />
 
           {selectedField && (
           <Autocomplete
